% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyEnsemble.R
\name{applyEnsemble}
\alias{applyEnsemble}
\title{Apply an ensembleModel to new data}
\usage{
applyEnsemble(
  ensembleModel,
  newDatabaseDetails,
  logSettings = PatientLevelPrediction::createLogSettings(),
  outputFolder
)
}
\arguments{
\item{ensembleModel}{An ensembleModel}

\item{newDatabaseDetails}{A databaseDetails object for the new database created using
\code{PatientLevelPrediction::createDatabaseDetails}}

\item{logSettings}{The log settings}

\item{outputFolder}{The location to save the base model results}
}
\description{
Apply an ensembleModel to new data
}
