% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyEnsemble.R
\name{applyEnsembleToPredictions}
\alias{applyEnsembleToPredictions}
\title{Apply an ensembleModel to list of base model prediction objects}
\usage{
applyEnsembleToPredictions(predictionList, ensemble)
}
\arguments{
\item{predictionList}{A list of base model prediction objects}

\item{ensemble}{An ensembleModel}
}
\description{
Apply an ensembleModel to list of base model prediction objects
}
