% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusionCombiner.R
\name{createFusionCombiner}
\alias{createFusionCombiner}
\title{Create the settings for a fusion ensemble}
\usage{
createFusionCombiner(
  type = c("uniform", "AUROC", "AUPRC")[1],
  evaluation = "CV",
  scaleFunction = "normalize"
)
}
\arguments{
\item{type}{The type of fusion ensemble pick from: 'uniform', 'AUROC', 'AUPRC' or any
other metric from evaluationSummary}

\item{evaluation}{The evaluation type used to learn the weights (if evaluation is CV and type
is 'AUROC' then the cross validation AUROC is used to determine the weight
given to of the base models)}

\item{scaleFunction}{How to scale the weights (normalize means weights add up to 1)}
}
\description{
Create the settings for a fusion ensemble
}
\examples{
\dontrun{

comSettingNormAUC <- createFusionCombiner(type = "AUROC",
                                          evaluation = "CV",
                                          scaleFunction = "normalize")

comSettingUniform <- createFusionCombiner(type = "uniform",
                                          evaluation = "CV",
                                          scaleFunction = "normalize")

}
}
