% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackerCombiner.R
\name{createStackerCombiner}
\alias{createStackerCombiner}
\title{Create the settings for a stacker ensemble - this is an emsemble that learns how to combine 
 level 1 models using labelled data}
\usage{
createStackerCombiner(
  levelTwoType = "logisticRegressionStacker",
  levelTwoHyperparameters = NULL,
  levelTwoDataSettings = list(type = "CV")
)
}
\arguments{
\item{levelTwoType}{The type of level 2 model (currently only supports "logisticRegressionStacker")}

\item{levelTwoHyperparameters}{The hyperparameter settings for the level 2 model}

\item{levelTwoDataSettings}{The settings specifying the data type to use to learn the level 2 model and the proportion of the data}
}
\description{
Create the settings for a stacker ensemble - this is an emsemble that learns how to combine 
 level 1 models using labelled data
}
\examples{
\dontrun{

stackerCombine <- createStackerCombiner(
levelTwoType = "logisticRegressionStacker",
levelTwoHyperparameters = NULL,
levelTwoDataSettings = list(type = 'Test', proportion = 0.5)
)


}
}
