% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblePatientLevelPrediction.R
\name{EnsemblePatientLevelPrediction}
\alias{EnsemblePatientLevelPrediction}
\title{EnsemblePatientLevelPrediction}
\description{
A package for developing ensembles using the PatientLevelPrediction framework

_PACKAGE
}
