% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterBaseModels}
\alias{filterBaseModels}
\title{Filter our base level 1 models that do not perform sufficiently well}
\usage{
filterBaseModels(resultList, filterSettings)
}
\arguments{
\item{resultList}{A list of runPlp() result objects (the level 1 models that are candidates for the ensemble level 1 models)}

\item{filterSettings}{A list containing the settings that specify how to filter the candidate level 1 models}
}
\description{
Filter our base level 1 models that do not perform sufficiently well
}
\examples{

\dontrun{

# only include models with an AUROC of 0.7 or higher for cross validation

level1Models <- filterBaseModels(
resultList, 
filterSettings = list(
evaluation = 'CV',
metric = 'AUROC',
minValue = 0.7
)
)

}
}
