% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoad.R
\name{loadEnsemble}
\alias{loadEnsemble}
\title{Load a previously saved ensemble result}
\usage{
loadEnsemble(dirPath)
}
\arguments{
\item{dirPath}{The directory to the saved ensemble result}
}
\description{
Load a previously saved ensemble result
}
\examples{
\dontrun{

ensemble <- loadEnsemble("C:/bestEnsemble")

}
}
