% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoad.R
\name{loadEnsembleModel}
\alias{loadEnsembleModel}
\title{Load a previously saved ensemble model (object of class ensembleModel)}
\usage{
loadEnsembleModel(dirPath)
}
\arguments{
\item{dirPath}{The directory containing the saved ensemble model}
}
\description{
Load a previously saved ensemble model (object of class ensembleModel)
}
\examples{
\dontrun{

ensembleModel <- loadEnsembleModel("C:/bestEnsembleModel")

}
}
