% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnsemble.R
\name{runEnsemble}
\alias{runEnsemble}
\title{Code to run the ensemble model development}
\usage{
runEnsemble(
  ensembleSettings,
  logSettings = PatientLevelPrediction::createLogSettings(logName = "ensemble"),
  saveDirectory,
  cohortDefinitions = NULL
)
}
\arguments{
\item{ensembleSettings}{The ensemble specifications created using \code{setEnsemble()}}

\item{logSettings}{The settings used to specify the logging, created using
\code{PatientLevelPrediction::createLogSettings()}}

\item{saveDirectory}{The location to save the ensemble}

\item{cohortDefinitions}{The cohort definitions for cohorts used in the ensemble}
}
\description{
Code to run the ensemble model development
}
