% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoad.R
\name{saveEnsemble}
\alias{saveEnsemble}
\title{Save an ensemble result}
\usage{
saveEnsemble(ensemble, dirPath)
}
\arguments{
\item{ensemble}{An ensemblePlp created by \code{EnsemblePatientLevelPrediction::runEnsemble()}}

\item{dirPath}{The directory to save the ensemble result}
}
\description{
Save an ensemble result
}
\examples{
\dontrun{

saveEnsemble(ensemble, "C:/bestEnsemble")

}
}
