% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoad.R
\name{saveEnsembleModel}
\alias{saveEnsembleModel}
\title{Save an ensemble model (object of class ensembleModel)}
\usage{
saveEnsembleModel(ensembleModel, dirPath)
}
\arguments{
\item{ensembleModel}{An ensembleModel}

\item{dirPath}{The directory to save the ensemble model}
}
\description{
Save an ensemble model (object of class ensembleModel)
}
\examples{
\dontrun{

saveEnsembleModel(ensembleModel, "C:/bestEnsembleModel")

}
}
