% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetEnsemble.R
\name{setEnsembleFromDesign}
\alias{setEnsembleFromDesign}
\title{Create setting for creating ensemble from model settings}
\usage{
setEnsembleFromDesign(
  modelDesignList,
  databaseDetails,
  filterSettings,
  combinerSettings
)
}
\arguments{
\item{modelDesignList}{A list of model designs to develop and then combine. Each model design is
created by \code{PatientLevelPrediction::createModelDesign()}}

\item{databaseDetails}{The OMOP CDM database details and connection for extracting the data}

\item{filterSettings}{Setting specifying rules to use to filter (remove) any model specified in
the list of model designs that performs insufficiently (these models get
ignored from the ensemble)}

\item{combinerSettings}{Settings specifying how to combine the remaining models into an ensemble}
}
\description{
Create setting for creating ensemble from model settings
}
\examples{
\dontrun{
library(FeatureExtraction)
library(PatientLevelPrediction)

covSet <- createCovariateSettings(
  useDemographicsGender = TRUE,  
  useDemographicsAgeGroup = TRUE
  )
  
restrictPlpDataSettings <- createRestrictPlpDataSettings()
  
modelDesign1 <- createModelDesign(
  targetId = 4,
  outcomeId = 3,
  restrictPlpDataSettings = restrictPlpDataSettings,
  covariateSettings = covSet,
  runCovariateSummary = F,
  modelSettings = setLassoLogisticRegression(),
  populationSettings = populationSet,
  preprocessSettings = createPreprocessSettings()
  )
modelDesign2 <- createModelDesign(
  targetId = 4,
  outcomeId = 3,
  restrictPlpDataSettings = restrictPlpDataSettings,
  covariateSettings = covSet,
  runCovariateSummary = F,
  modelSettings = setGradientBoostingMachine(),
  populationSettings = populationSet,
  preprocessSettings = createPreprocessSettings()
  )

ensembleSettings <- setEnsembleFromDesign(
  modelDesignList = list(modelDesign1, modelDesign2),
  databaseDetails = createDatabaseDetails(),
  filterSettings = list(minValue = 0.5, maxValue = 1),
  combinerSettings = createFusionCombiner(
    type = "uniform",
    evaluation = "CV",
    scaleFunction = "normalize"
    )
  )
}
}
