% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnsemble.R
\name{runEnsemble}
\alias{runEnsemble}
\title{Code to run the ensemble model developement}
\usage{
runEnsemble(
  ensembleSettings,
  logSettings = PatientLevelPrediction::createLogSettings(logName = "ensemble"),
  saveDirectory
)
}
\arguments{
\item{ensembleSettings}{The ensemble specifications created using \code{setEnsemble()}}

\item{logSettings}{The settings used to specify the logging, created using \code{PatientLevelPrediction::createLogSettings()}}

\item{saveDirectory}{The location to save the ensemble}
}
\description{
Code to run the ensemble model developement
}
